Feat. replace useQuery with hook and fix refetch problem when creating new doc
