+++ src/components/HeaderItems/DocHandlerButtons.jsx
@@ -1,8 +1,6 @@
-import { useQueryClient } from "@tanstack/react-query";
 import { useAtom, useAtomValue, useSetAtom } from "jotai";
 
 import {
-  currentDBIdAtom,
   currentDocIndexAtom,
   isEditModeAtom,
   documentsIdsAtom,
@@ -12,15 +10,12 @@ import {
   documentsNumAtom,
 } from "../../atoms/atoms";
 
-import useGetAllDocuments from "../../apis/useGetAllDocuments";
-
 import Button from "../shared/Button";
 import AddDocModal from "../Modals/AddNewDocument/AddDocModal";
 import DeleteDocModal from "../Modals/DeleteDocument/DeleteDocModal";
 
 function DocHandlerButtons() {
   const documentsNum = useAtomValue(documentsNumAtom);
-  const queryClient = useQueryClient();
 
   const [currentDocIndex, setCurrentDocIndex] = useAtom(currentDocIndexAtom);
   const [showAddDocumentModal, setShowAddDocumentModal] = useAtom(
@@ -30,7 +25,6 @@ function DocHandlerButtons() {
     showDeleteDocumentModalAtom,
   );
 
-  const currentDBId = useAtomValue(currentDBIdAtom);
   const isEditMode = useAtomValue(isEditModeAtom);
   const documentsIds = useAtomValue(documentsIdsAtom);
 
@@ -38,27 +32,15 @@ function DocHandlerButtons() {
 
   const currentDocIndexShownToUser = currentDocIndex + 1;
 
-  useGetAllDocuments();
-
   function navigateDown() {
     if (currentDocIndexShownToUser !== 1) {
       setCurrentDocIndex(prev => prev - 1);
-      queryClient.refetchQueries([
-        "foreignDocuments",
-        currentDBId,
-        currentDocIndex,
-      ]);
     }
   }
 
   function navigateUp() {
     if (currentDocIndexShownToUser !== documentsNum) {
       setCurrentDocIndex(prev => prev + 1);
-      queryClient.refetchQueries([
-        "foreignDocuments",
-        currentDBId,
-        currentDocIndex,
-      ]);
     }
   }
